Cleanup new GDB code
